MCOL-4394 Followup

+++ datatypes/mcs_int128.h
@@ -185,7 +185,7 @@ class TSInt128
       return toDouble();
     }
 
-    inline long double toDouble() const
+    inline double toDouble() const
     {
       return static_cast<double>(s128Value);
     }
